requirements.txt
cs50
Flask
Flask-Session
werkzeug.security
functools
requests